Fix program loading
One problem was that there are many `SHT_PROGBITS` besides what is interesting
to us, like `.text` `.eh_frame`, and others. We need to be sure that we're
loading only programs that we're interested in.

The other is that instead of an index, the `sh_name` contains an offset, so
the string table needs to be accessed differently.

This commit is a bit messy, should clean up later.