Rework stats aggregation
Move the statsd aggregation code into its own, separate aggregator
actor. With this change the statsd probe becomes a light statsd protocol
parser and the aggregation logic becomes usable with data coming from
other probes.

Counters and gauges are aggregated as you would expect, sets are
aggregated as a single SET_UNIQUES output metric and histograms are
aggregated as a list of output percentiles. The percentiles are
currently hardcoded to 25, 50, 75, 90, 95, 99. In the future they'll be
configurable through config.toml.