ci: rework cargo_build
Add condition: <shell condition> parameter that can be used to skip
builds. This is now used by the ubuntu and fedora targets to skip
release builds unless AWS_EC2_SSH_KEY_ID is set, which means that the
workflow is being run in the official repo, and that terraform is going
to need the release binaries.